Spec/hhbbc change: AsyncESuspend only captures (and unsets) named locals
It only copies named locals into the suspended async
function.  The behavior looks like it dates back to continuations, and
may need to change?  I couldn't think of a case where we currently
would emit an AsyncESuspend while an unnamed local is still needed,
but it hit some assertions for type-inference if you suspended in a
string switch (uses an unnamed local that's live across the cases).  I
suppose if nothing else with a string switch you could observe that
the destructor of an object ran earlier iff the function was
suspended, and if we use unnamed locals for more things this could
bite us later ...

Reviewed By: @dariorussi

Differential Revision: D1166848