Disable /Zend/tests/inter_06.php in repo mode
Output from HPHPc includes classes with non-unique names that
have AttrUnique if you declare interfaces with the same names as
builtins.  This only causes a problem because hhbbc also has its own
code to determine which classes are unique now, and when it doesn't
match what hphpc said it asserts.  I would rather keep the asserts
than keep this test for now---when we are ready to remove the hphpc
code that detects class uniqueness this will work in hhbbc.

This presumably relates to #3075835.

Reviewed By: @swtaarrs

Differential Revision: D1167199