Add Code Model classes to systemlib
The proximate reason for the move is to enable PHINQ to work every that HHVM gets deployed to. However, it is also anticipated that people may end up writing scripts that parse a PHP file into a Code Model and then traverse the model to look for various syntactic patterns. The latter scenario is enabled by moving fb_serialize_code_model_for to the public part of systemlib and renaming it to get_code_model_for.

Although this is little more than a code move, the generator scripts had to be adapted to place the generated code into a namespace (HH\CodeModel) so that there is less chance of name clashes arising in existing code. The introduction of a namespace, in conjunction with a number of bugs relating to namespaces, made it necessary for the generator scripts to comment out type annotations.

Despite appearances, this diff is actually a pretty minor change.

Reviewed By: @paroski

Differential Revision: D1165337