Emit mergeOnly units so their pseudo-main still returns what it should
Right now for mergeOnly units, we set a TypedValue in
metadata for what the pseudo-main should return, but emit a
pseudo-main that returns 1.  This is causing an issue for hhbbc if you
want to re-examine the bytecode.  (We could teach it to read from the
mainReturn TypedValue, but that seems more complicated.)

It is working when you don't do hhbbc as a second independent step,
because Option::UseHHBBC turns off all the merge only logic in the
emitter (except for systemlib).

Reviewed By: @jasone

Differential Revision: D1171376