Enable "FuncFamily" option by default
Over several runs its been noise-to-maybe-slightly good in
perflab.  It's not entirely clear what was making it a regression
initially (although assert type filtering must've helped some).  The
only actually green metric when you turn this on is branches/branch
misses.  Also rename the option, since it reads weird for a -ffoo flag
to say "enable" or "no-enable".

Reviewed By: @dariorussi

Differential Revision: D1171581