Simple local DCE for hhbbc
I wanted to make sure backward algorithms that needed access
to types will work somehow, so I figured trying this was a way to
check how it might turn out.  It seems like it's not a big deal* to
just propagate the states we got from the forward analysis to each
instruction, so we can see them while going backwards.  This prototype
catches very little dead code in www, however.  I think implementing
global DCE so we can remove stores to locals might find a bit more
(and probably would use this as part of it), but this is probably
enough to know this isn't likely to matter very much yet.

Also, this seems like it's really time to split up
abstract-interp.cpp.  All this code should be in its own file.  (I'll
do that soon.)

*: I was worried about compile-time perf, but probably in part due to
@dariorussi taking PropStates out of there it is apparently very
reasonable to copy States everywhere.  (This pass only adds 3seconds
in opt to the time taken building www.)

Reviewed By: @dariorussi

Differential Revision: D1174688