Allow join expressions to be less verbose.
When joining entity tables with association tables, it is obvious which columns have to match for two tuples to join up. The C# syntax for joins is too verbose in such situations. The syntax is now enhanced to make the on condition optional. The table alias is now also optional, since there can now be situations where no explicit references exist to the columns of a table.

Reviewed By: @paroski

Differential Revision: D1169047