Support bytecode control flow in IR regions
Adds the ability to generate IR representing bytecode-level control flow.  IRBuilder now keeps a map of bytecode offsets to IR blocks, which we add to when translating a JmpZ/JmpNZ in which both paths are in the region.  Before translating a bytecode block, the map is checked to see if state (local types) needs to be merged, and if the merged states are compatible.  Incompatible merges force the translator to create an exit on one of the paths; otherwise a normal control flow merge point is created.

Caveats: this has only been tested with hottrace regions, so the design is based around creating "branchovers" (a forward jump to a block later in the trace).  Loops are unhandled, and divergent control flow may be buggy.  Also, this doesn't work with refcount opts or the extra simplifier pass.  Those passes share the state-management code with IRBuilder, and some additional work is needed to support both.

Reviewed By: @edwinsmith

Differential Revision: D1154657