Tweak Type to make it easier to add data with non-trivial copy constructors
I think we're probably going to want to put things in Data
that have non-trivial copy constructors at some point (e.g. for
non-static array subtypes), and more than once I've wanted to do it
temporarily as part of prototyping a type system change.  Using
folly::Optional<> makes it hard for this case, because Data is a union
and would have to have a copy constructor, but it doesn't know how to
copy itself because it doesn't contain its own type tag.

Reviewed By: @swtaarrs

Differential Revision: D1177979