Handle non-trivial Type::Data copying, put the wait handle type into DObj
It seems more natural in here.  It should probably get its
own DataTag later, but for now this keeps dobj_of, is_specialized_obj,
and a few other things working with fewer special cases (which would
involve carrying around a pointer to the Index if it has a wait handle
to be able to get a res::Class for WaitHandle later).

This also adds a separate tag for specialized data.  This should make
things a little easier when adding new types that can have data,
because we can have exhaustive switch statements.  It also means we
can have some type bit patterns that could support multiple types of
data later (I will probably need this for arrays).

Reviewed By: @swtaarrs

Differential Revision: D1178029