Release Elf object after extracting embedded data
elf_begin mmap's the file it's given, so we were ending up
with multiple copies of the hhvm binary loaded whenever we made
multiple calls to get_embedded_data, leading to a huge amount of
virtual memory allocated.

This was causing several CLI-mode flib tests to fail when run with a
finite ulimit.

Reviewed By: @alexmalyshev

Differential Revision: D1179488