remove operator-> from String
Basically i removed the operator, then opened all the files with compilation errors and either used a method if it already existed or used `.get()` if i thought it was seldomly used.

We can do another pass after and replace all the `.get()->foo()` with `.foo()` if `foo()` is really popular. (`isame` seems plausible).

Reviewed By: @edwinsmith

Differential Revision: D1180662