make toCppString consistent casing
I just picked one, do you prefer the other casing?

I was really doing this because I made a mistake and did `->toCPPString` on a `String` instead of `.toCppString` which deals with null strings. There were a few other places in the code which look like they did the same thing.

Reviewed By: @edwinsmith

Differential Revision: D1180567