Add comprehensive test for PHINQ
This test consists of a fully featured query provider for SQLite3 plus a number of query expressions that cover most of the logic. As a consequence of doing the query provider and test cases a number of bugs came to light and it became clear how to lift various restrictions. Some highlights:
1) Instead of using the arguments to function calls to decide if they should happen client-side or server-side (which breaks down for constant arguments), there is now the convention that server-side function names are prefixed with q_.
2) Any expression in a select clause that can be evaluated server-side is now retained and sent to the server.
3) The consequent changes to the select clause call back closure brought to light a problem with how the closures were constructed, which manifested itself as silent optimization failures in repo mode. This diff hopefully fixes that.

Reviewed By: @paroski

Differential Revision: D1184868