Remove incorrect assertion about proflogues
This assertion fires regularly in debug builds, because it's
not true.  It asserts that, during regeneration of PGO prologues, if a
translation has a generated prologue (other than the default unique
stub) that it also has an entry in the profile data.  This can broken
because we discover new candidate functions for profiling during
interpretation.  The sequence is: (1) a prologue for a non-AttrHot
function is generated; (2) meanwhile, that function executes in the
interpreter enough times to be marked AttrHot; (3) upon retranslation,
this function has a translated prologue but no profData entry.

There's not really a good way to get this right (you could try to
regenerate proflogues upon AttrHot promotion, but that's dependent on
having the write lease, so could still fail), but fortunately there's
no correctness issue here - we'll just sometimes retranslate with less
profile data than we might have liked.

Reviewed By: @swtaarrs

Differential Revision: D1182212