Fix a bug with local bases of static string types
In a later change I've change it so SStr=" " couldBe SStr=""
returns false.  Previously all static strings said they couldBe the
empty string, so elemCouldPromoteToArr would be true and we'd turn the
local into an InitCell (union of SStr and Arr), so we were doing
things "right" without handling this case explicitly.

Reviewed By: @dariorussi

Differential Revision: D1186440