Hook specialized array types for locals up to member instructions
Adds support in the interpreter for creating specialized
array types in NewStructArray, NewPackedArray, or through vector
instructions affecting local types.  Arrays in properties are going to
be a little different because they are control flow insensitive, so
it's not hooked up yet, but it will probably involve a similar "array
chain" thing in a different mode.

Reviewed By: @dariorussi

Differential Revision: D1196144