Introduce state widening operations to deal with infinitely growing types
After the addition of array inner types, we have an issue
that union_of is no longer guaranteed to terminate in finite
steps---it is possible to make continuously larger array types without
reaching a fixed point.

The solution to this is to introduce a stronger union operation
(widening) that is guaranteed to take only finite applications to
reach a stable point, and then use it occasionally if we think we're
iterating too much.

Similarly, we now will be able to have infinitely shrinking chains in
return types in the index.  The index is conceptually easier to fix:
since it's never allowed to contain anything incorrect, it's never
wrong to stop refining its types.

Reviewed By: @edwinsmith

Differential Revision: D1186766