only register ini settings at threadInit instead of requestInit
This is similar to how zend does it. Basically you you bind your ini settings to thread-local memory right when the thread starts up, then on any future `ini_set` we do an `ini_get` first, save the value and then do the set. Then when the request ends we replay all the saved values to put the thread back to its state before the request ran.

Reviewed By: @fredemmott

Differential Revision: D1186887