Remove Type::None
This is all pretty mechanical after removing jump dests. I had to
replace None with Bottom in two places: the default constructor of Type
(which seems like it shouldn't exist but I'd rather try to deal with in
another diff), and the output type of a DMulti instruction. DMulti is
only used for DefLabel, which will just have placeholder dests until it
receives incoming edges. The Bottom dests will disappear before the end
of initial IR generation, so I think it's not that bad.

Reviewed By: @swtaarrs

Differential Revision: D1189637