hhvm manual TOC, some basic jit docs
I had a lot of trouble trying to decide what granularity we want in
the TOC, but it's best to just get something in and then iterate as we fill it
out. This also includes some docs on region selection and a couple
optimizations. It's written for the (hopefully) near future where
Translator::analyze has been replaced by the region selector.

Reviewed By: @ottoni

Differential Revision: D1191045