IniSetting::Bind is now typed
instead of the weird `void* p` we now have the slightly less weird `IniSetting::SetAndGet<bool>`. Is this easier to read? I think it is nicer to code from scratch and will help me replace hdf files.

I basically switched from saving the `p` in the callbacks structure to saving it in the captured lamdbas.

Reviewed By: @fredemmott

Differential Revision: D1191614