Raise warnings for creating default object after modifying the base
Zend appears to actually allow the entire expressions's
effects to occur before it raises warnings for creating default
objects.  We don't really want to do this, but this matches slightly
closer, and has the advantage that it doesn't propagate the Uninit
state to the catch block (which would let Uninit fall through after
the catch).  This also matches closer with implicit array creation
now.

Reviewed By: @paroski

Differential Revision: D1195860