Add toImmMap() and immutable() methods to collections
This diff adds toImmMap() to the KeyedIterable interface. toImmMap() is a
materialization method like toMap() and toImmVector().

I also added an immutable() method to each of the concrete collection types
to allow the programmer to get an immutable copy of the collection. Note
that calling immutable() on an immutable collection (such as ImmVector)
will simply return itself.

Reviewed By: @elgenie

Differential Revision: D1188623