Allow resetting the default statement class to PDOStatement
Currently HHVM strictly enforces that the statement class be a descendent of PDOStatement. While the PHP Documentation states that it has to be a subclass, I believe this is subclass in the sense implied by Liskov's OO substitution principle. Thus, the PDOStatement class should be accepted.

Zend does accept the PDOStatement class without warning or notice.

Closes #1951
Closes #1954

Reviewed By: @fredemmott

Differential Revision: D1210794