Rename util.h to file-util.h to reflect its smaller intent.
Removed references to util.h and added back just the
minimal includes of file-util.h.  Also changed the last
few lingering mentions of the Util namespace to FileUtil,
since removing the namespace entirely would require
new :: qualifiers on some of the posix functions that
collide with names in FileUtil, like mkdir() and rename().

Reviewed By: @jdelong

Differential Revision: D1192454