Fix some tracelet region selector stuff
When an inlined FCall has its output marked as predicted, analyze
handles it as though the call weren't inlined, so it ends up predicting the
output type of the inlined call. Since NormalizedInstructions aren't persistent
in selectTracelet, keep track of the return type prediction, if any, when
entering an inlined call.

Get rid of the unused startSk argument from outputIsPredicted.

Reviewed By: @ottoni

Differential Revision: D1199827