Instruction, load and store counters for vixl
We'd like to be able to track the progress of ARM codegen, so here are
some counters to expose in Trace::stats.

We're also going to need something to track interped vs. translated
bytecodes, but I think that's going to require more finesse, so I'll do
it separately.

Just do ##TRACE=stats:1 hhvm --arm whatever.php## and at the end of
execution it will dump all the stats counters.

Reviewed By: @edwinsmith

Differential Revision: D1194779