Fix assertion in memory-manager.cpp
It turns out that a jemalloc sync can occur between the initialization of the MemoryManager and when refreshExternalStats is called during hphp_session_init to erase the record of shared structure initialization. This breaks multiple assumptions, so disallow syncs until we really do want to start tracking. This refines the related assertions as well.

Reviewed By: @edwinsmith

Differential Revision: D1198787