Fix an overflow issue with Vector::slice()
In some cases where the 'len' parameter was really big (ex. PHP_INT_MAX),
the math performed by Vector::slice() could overflow and cause HHVM to
crash. This diff fixes the math so that it doesn't overflow. It also cleans
up the implementation and gets rid of ImmVector::slice() (since the static
slice() methods are deprecated).

Reviewed By: @elgenie

Differential Revision: D1203723