Clean the list of shutdown handlers even if they fail
If a shutdown handler fails (fatal, time/memory limit exceeded, etc.),
the list of shutdown handlers is not cleared.

This is normally not a big deal, since shutdown handlers are invoked as
the last thing before destroying ExecutionContext. However, in the RPC
case, the same ExecutionContext is reused across multiple RPC requests
and the same failed shutdown handler is then executed again after the
next (unrelated) RPC request. This previously failed shutdown handler is
likely to fail again and keep failing until ExecutionContext is
reinitialized.

Reviewed By: @markw65

Differential Revision: D1202917