In TransProfile mode, only do HHIR guard relaxation with tracelet region selector
In TransProfile mode, we can only relax the guards in tracelet region
mode.  If the region came from analyze() and we relax the guards here,
then the RegionDesc's TypePreds in ProfData won't accurately reflect
the generated guards.  This can result in a TransOptimze region to be
formed with types that are incompatible, e.g.:

    B1: TypePred: Loc0: Bool      // but this gets relaxed to Uncounted
        PostCond: Loc0: Uncounted // post-conds are accurate
    B2: TypePred: Loc0: Int       // this will always fail

Reviewed By: @swtaarrs

Differential Revision: D1197001