Deal with empty arrays better
Previously, all the specialized array types implied that the
array was non-empty.  This fails to get types in some common idioms
(e.g. populating a vector-like array in a loop, where it was empty
going into the loop.)  This diff just adds new bits for empty arrays,
so any of the specialized array types can be unioned with either
counted or uncounted empty arrays.

Reviewed By: @dariorussi

Differential Revision: D1208262