change the debugger to use .ini files
In my quest to kill `.hdf` files these have to go. The weirdest thing was writing out all the colors to the file just incase someone wrote a node to reference them, instead I just put them in the docs.

There isn't an INI serializer built into php, but the format was simple enough to do with strings.

Reviewed By: @mikemag

Differential Revision: D1173217