Fix a bug in CGetL; FPushObjMethodD grouping
This can cause Index types to grow.  The first time through,
we might decide a function returns TObj because we had a TInitCell
base on a FPushObjMethodD.  The next time through, if we have a
specialized object type that is optional, we'll leave it as OptObj,
which isn't a subtype of what we inferred last time so it breaks index
invariants.

Reviewed By: @dariorussi

Differential Revision: D1209348