Break the circular dependency between IRBuilder and Simplifier
The dependency chain that encompasses how Simplifier gets its results
into IRBuilder looked like this:

  optimizeInst -> optimizeWork -> Simplifier::simplify ->
  Simplifier::gen -> IRBuilder::gen -> optimizeInst

This is suboptimal if you're trying to tease out parts of Simplifier so
that it can be unit-tested. This diff breaks that dependency.

The main point is to reformulate the interface to Simplifier::simplify
so that it reflects what the result of simplification actually is: zero
or more instructions, along with an SSATmp to use in place of the
simplified instruction's dst. This way, Simplifier can return multiple
instructions without needing the backdoor afforded by IRBuilder::gen.

The new setup is that Simplifier recursively calls itself directly.  One
notable consequence is that we no longer do CSE interspersed with
simplification (since IRBuilder::optimizeWork does CSE); instead, we do
all simplification up front and then do CSE on the result. I think this
should produce the same output, and may do so with fewer CSE lookups.

Reviewed By: @swtaarrs

Differential Revision: D1219292