Change ext_xml's use of Variant::set to use Array::set
parser->ctag is strong bound to the last element in
parser->data, and parser->data is only manipulated by calling append
with arrays, or by assignRefing it to an array created in
xml_parse_into_struct, so ctag must always be an array.  Similar story
for curTag but it has a smaller scope so it's more obvious.

Reviewed By: @markw65

Differential Revision: D1213703