Remove Variant::set functions from IniSetting
I'm not 100% sure what's going on in onEntry.  It seems like
the arg should always be an array, but I'll ask @ptarjan.  (For now
I'm converting it if it isn't.)  I made CallbackData private to be
sure I wasn't missing creation sites (but it's only used in this class
and its member classes).

Reviewed By: @ptarjan

Differential Revision: D1213717