Eliminate m_entryFP from ExecutionContext::m_nestedVMs
ExecutionContext::m_nestedVMs is a vector of saved state of nested
VMs (excluding the one that is currently running). This state contains
firstAR field pointing to the ActRec the corresponding VM was entered
at. In addition, m_nestedVMs contains m_entryFP pointing to the ActRec
of the VM we entered into at the time the m_nestedVM record was created.
This is redundant to the firstAR field (the data are shifted by one
index) and this diff eliminates it.

Reviewed By: @paroski

Differential Revision: D1214373