Add __IsFoldable user attribute for consty expressions
HNI/user version of IDL IsFoldable flag.

I had initially considered only setting isFoldable
on the function scope, but this ran into some ordering
issues (which weren't a problem for IDL builtins).
This approach also has the benefit of saving the foldability
to the repo instead of losing it post-compile.

Along the way I split parseUserAttributes and parseNativeAttributes
into separate helpers.  User is always called (for UAs which
apply to both HNI and PHP functions), while Native is
only caled on HNI functions.

Mark hash(), hash_hmac(), and hash_algos() as foldable
to exercise this API.

Reviewed By: @dariorussi

Differential Revision: D1212999