accept file which has only <?php directive, followed by EOF (Zend compat)
Zend will accept a file that has only <?php directive in it (with no further whitespace or newline.  HHVM gave a parse error.   The issue is in the lexer.   Unfortunately, detecting EOF in flex seems to be a little annoying, because while there is <<EOF>> pattern, it is only valid to use it by itself, not as part of a more complex pattern.   The rule in question is on line 733 in hphp.ll.  We want to make sure that <?php is followed by whitespace or a newline, i.e. not <?phpsomeothercrap...  There is no good way to put the EOF test here.

My proposal is to have add a new rule for just "<?php".  Since the original rule on 733 will match a longer text (including the whitespace or newline), it will get the priority in most cases.  If we match just '<?php', we enter a new state ST_IN_PHP_OPEN_TAG, where it is not valid to have any further symbols.  Since the EOF is not a separate symbol in lex, the empty file will pass and not hit the rule on L763.  L763 will cause an error in all other cases.

If this seems too convoluted, another approach to solving this is just to append a new line to any source text before attempting to lex it.  This is smaller fix, but seems kinda hacky.

Zend has a similar special handling, look for "inline_char_handler:" in zend_language_scanner.l.

Reviewed By: @ptarjan

Differential Revision: D1215503