Fix crash in ArrayObject
The call to ElemDObject here was incorrectly forwarding keyType. The
call to initScratchKey at the beginning of the function modifies key to point
to a real TypedValue, so any subsequent uses of key need to use KeyType::Any.

This would've been prevented if we were using a proper Key type instead of
abusing TypedValue*, which is something I tried to do a while ago. It
inexplicably caused a perf regression at the time so I abandoned it, but that
was with an old gcc so we should try again.

Closes #1948

Reviewed By: @bertmaher

Differential Revision: D1221410