Fix a bug in ldClsPropAddr
When ldClsPropAddr encounters a class that is the same name
as the context class, but the context class is not loaded yet (which
is possible if you're inlining a static call), it can hit a code path
that never happened before where the clsVal wasn't constant but
useSpropCache was true.  Fix this.

Reviewed By: @edwinsmith

Differential Revision: D1223040