Genericize HHBBC::SStringOr into a Either<L,R> class
This is a fairly common idiom.  I thought I might use it in
FullPos, so I genericized it, but I'm not 100% sure if I'll use it
there yet.  (Also, uses in things like iterators probably would
justify an API that gets the thing without a type check.)  Seemed
worth committing for now anyway.

Reviewed By: @alexmalyshev

Differential Revision: D1223461