Remove m_strongIterators from ArrayData; use a table on the side in TLS
This is very rare, so we do better putting it on the side.

The logic for the thread local table is a little bit unusual, because
I wanted to keep it as easy as a single compare to determine if an
array might need to update or free strong iterators.  Also one thing
about this scheme that seemed a little unfortunate at first is that it
slows down all array mutations on arrays of kinds supporting strong
iteration if any arrays in the program are being strong iterated.

I tried to fix both of those things of those by using a bit in the
array header that gets set if an array is ever strong iterated.  But
that increased dcache and dTLB misses compared to this version---a
hybrid (that checks TLS first and then checks the array header next)
might do better, but I think now the better route to that might just
be to eventually have more of the array kinds imply an array has never
been strong iterated, so fewer arrays will eventually be affected by
strong_iterators_exist() returning true.

Reviewed By: @edwinsmith

Differential Revision: D1225581