Assert property repoAuthTypes in debug builds
This asserts the types hhbbc has inferred for properties
whenever they are read from, if it's a debug build.  It caught that
I'm leaving Closure use var properties as Uninit (harmless, but
technically not so cool).  We could also assert them on sets, but for
now just doing gets.  (Sets are slightly more painful, since the
places that actually do the set are a lot more spread out.)

Reviewed By: @dariorussi

Differential Revision: D1226026