Add exception barriers before jumps to retain spillstacks
This fixes a subtle bug that was breaking a flib test.  The bug manifested like this:

- before jumping to a merge point, the stack pointer has to be sync'ed using a SpillStack;
- at a merge point we insert a DefSP to prevent any future instructions from walking the SP chain back into one of the conditionals;
- because the DefSP redefines the SP, the earlier SpillStack was being DCE'ed.

The solution here is to just put an exception barrier after any spill associated with a jump.  At first, this seems kind of heavy handed,
but (a) it's conditional on using HHBC control flow, and (b) SpillStacks don't really tend to get DCE'ed anyways.

Reviewed By: @edwinsmith

Differential Revision: D1221048