Fix Autocomplete of ID
D1195838 broke this. (My fault, I need to make a test harness and write some tests for this stuff)

The problem is when you're starting to type a global, it doesn't necessarily hit the dispatch_call part.

(There's no brackets typed yet)

This moves that code out of fun_type_of_id and adds it in both those call sites and one more.

Reviewed By: @jwatzman

Differential Revision: D1230498