Unit test for AbsDbl and jump fusion
Well, this is what it looks like. I did a few basic tests of jump
fusion because I've been going after that for ages and ages.

I had to create a new test-only opcode, which I called Conjure. It's
basically non-const DefConst; its purpose is to produce a value
of a given type that no optimization can see through.

Everyone I've talked to about my plans for unit testing the JIT has
mentioned needing some sort of assembler for HHIR, but I've always been
a little confused as to why that's needed. As this diff demonstrates,
it's entirely possible to create IR "in the laboratory" without anything
like an assembler. Is the way it's done here really that bad? Admittedly
the way I verify the output is a little clunky, but I don't immediately
see how having an IR assembler would make it significantly better.

Reviewed By: @bertmaher

Differential Revision: D1234437